+++ chartlets.js/src/lib/components/Checkbox.tsx
@@ -3,14 +3,18 @@ import MuiCheckbox from "@mui/material/Checkbox";
 import MuiFormControl from "@mui/material/FormControl";
 import MuiFormControlLabel from "@mui/material/FormControlLabel";
 
-import { type CheckboxState } from "@/lib/types/state/component";
-import { type ComponentChangeHandler } from "@/lib/types/state/event";
+import { type ComponentState } from "@/lib/types/state/component";
+import type { ComponentProps } from "@/lib/component/Component";
 
-export interface CheckboxProps extends Omit<CheckboxState, "type"> {
-  onChange: ComponentChangeHandler;
+interface CheckboxState extends ComponentState {
+  label?: string;
+  value?: boolean | undefined;
 }
 
+interface CheckboxProps extends ComponentProps, CheckboxState {}
+
 export function Checkbox({
+  type,
   id,
   name,
   value,
@@ -22,7 +26,7 @@ export function Checkbox({
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     if (id) {
       return onChange({
-        componentType: "Checkbox",
+        componentType: type,
         id: id,
         property: "value",
         value: event.currentTarget.checked,