+++ chartlets.js/src/lib/types/state/event.ts
@@ -1,8 +1,7 @@
-import { type ComponentType } from "@/lib/types/state/component";
 import type { ObjPathLike } from "@/lib/utils/objPath";
 
 export interface ComponentChangeEvent {
-  componentType: ComponentType;
+  componentType: string;
   // See commonality with StateChange
   id: string;
   property: ObjPathLike;