+++ chartlets.py/chartlets/components/__init__.py
@@ -1,5 +1,6 @@
 from .box import Box
 from .button import Button
+from .button import IconButton
 from .checkbox import Checkbox
 from .progress import CircularProgress
 from .progress import CircularProgressWithLabel